+++ Server/src/GameWorld/Collision/Collidable.cpp
@@ -77,7 +77,7 @@ bool Collidable::isAlignedLeft(const std::shared_ptr<Collidable> &other) const {
     return (not this->isOnRight(other) and (this->overlapVerticalAxis(other)));
 }
 #include <iostream>
-bool Collidable::isCloseTo(const std::shared_ptr<Collidable> &other, const float& closeDistance) const {
+bool Collidable::isCloseTo(const std::shared_ptr<Collidable> &other, const double& closeDistance) const {
     /*std::cout << "\nCOLLIDABLE:\nInfected Pos: x:" + std::to_string(this->position.first) + " y: " +
                  std::to_string(this->position.second) << std::endl;
     std::cout << "Grenade Pos: x:" + std::to_string(other->position.first) + " y: " +
